+++ backend-api/backend-api/Features/TaskForum/Controller/ChannelController.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using backend_api.Database.ProjectRepository;
+using backend_api.Database.ProjectTaskRepository;
 using backend_api.Features.TaskForum.Models;
 using backend_api.Features.TaskForum.Repository;
 using backend_api.Util;
@@ -10,21 +12,18 @@ namespace backend_api.Features.TaskForum.Controller
     [Route("api/[controller]")]
     public class ChannelController : Microsoft.AspNetCore.Mvc.Controller
     {
-
         private readonly IChannelRepository _channelRepository;
 
         public ChannelController()
         {
             _channelRepository = new ChannelRepository();
         }
-        
+
         // CREATE
         // POST api/channel/:taskId
         [HttpPost("{taskId}")]
-        public ActionResult Post(string taskId, [FromBody]Channel channelToCreate, [FromHeader]string userId)
+        public ActionResult Post(string taskId, [FromBody] Channel channelToCreate, [FromHeader] string userId)
         {
-            // TODO: Check valid auth and access
-            
             // Parse task guid and user guid
             var taskGuid = ValidationUtil.ParseGuid(taskId);
             var userGuid = ValidationUtil.ParseGuid(userId);
@@ -32,29 +31,38 @@ namespace backend_api.Features.TaskForum.Controller
             {
                 return BadRequest(Errors.InvalidGuid);
             }
-            
-            // TODO check task exists
-            
+
+            // Check task exists
+            var existError = CheckTaskExists(taskGuid);
+            if (existError != null)
+            {
+                return existError;
+            }
+
+            // Check user is an owner of the project so has permission to access
+            var error = CheckGroupMember(taskGuid, userGuid);
+            if (error != null)
+            {
+                return error;
+            }
+
             // Check valid channel name
             if (string.IsNullOrEmpty(channelToCreate.Name.Trim()))
             {
                 return StatusCode(400, Errors.ChannelNameEmpty);
             }
-            
+
             var result = _channelRepository.Add(channelToCreate, taskGuid);
-            
-            if (result.IsError)
-            {
-                return StatusCode(500, result.ErrorException.Message);
-            }
-            
-            return StatusCode(201, result.ReturnValue);
+
+            return result.IsError
+                ? StatusCode(500, result.ErrorException.Message)
+                : StatusCode(201, result.ReturnValue);
         }
 
         // READ
         // GET api/channel/:taskId
         [HttpGet("{taskId}")]
-        public ActionResult<List<Channel>> GetAll(string taskId, [FromHeader]string userId)
+        public ActionResult<List<Channel>> GetAll(string taskId, [FromHeader] string userId)
         {
             // Parse task guid and user guid
             var taskGuid = ValidationUtil.ParseGuid(taskId);
@@ -63,27 +71,36 @@ namespace backend_api.Features.TaskForum.Controller
             {
                 return BadRequest(Errors.InvalidGuid);
             }
-            // TODO
-            // Check auth/user - Unauthorised() and Forbidden() if no access to project
-            // Check task exists - NotFound()
-            
+
+            // Check task exists
+            var existError = CheckTaskExists(taskGuid);
+            if (existError != null)
+            {
+                return existError;
+            }
+
+            // Check user is an owner of the project so has permission to access
+            var error = CheckGroupMember(taskGuid, userGuid);
+            if (error != null)
+            {
+                return error;
+            }
+
             var result = _channelRepository.GetAllForTask(taskGuid);
-            
+
             if (result.IsError)
             {
                 return StatusCode(500, result.ErrorException.Message);
             }
-            
+
             return StatusCode(200, result.ReturnValue);
         }
 
         // UPDATE
         // PATCH api/channel/:channelId
         [HttpPatch("{channelId}")]
-        public ActionResult Patch(string channelId, [FromBody]Channel channelToUpdate, [FromHeader]string userId)
+        public ActionResult Patch(string channelId, [FromBody] Channel channelToUpdate, [FromHeader] string userId)
         {
-            // TODO: Check User auth and access - Unauthorised() || Forbidden()
-            
             // Parse channel guid and user guid
             var channelGuid = ValidationUtil.ParseGuid(channelId);
             var userGuid = ValidationUtil.ParseGuid(userId);
@@ -91,35 +108,44 @@ namespace backend_api.Features.TaskForum.Controller
             {
                 return BadRequest(Errors.InvalidGuid);
             }
+
             channelToUpdate.Guid = channelGuid;
-            
-            // TODO: Check channel exists
-            
+
+            // Check channel exists
+            var existError = CheckChannelExists(channelGuid);
+            if (existError != null)
+            {
+                return existError;
+            }
+
+            // Check user is an owner of the project so has permission to access
+            var task = _channelRepository.GetTask(channelGuid);
+            var error = CheckGroupMember(task.ReturnValue.Guid, userGuid);
+            if (error != null)
+            {
+                return error;
+            }
+
             // Check valid channel name
             if (string.IsNullOrEmpty(channelToUpdate.Name.Trim()))
             {
                 return StatusCode(400, Errors.ChannelNameEmpty);
             }
-            
+
             var result = _channelRepository.Edit(channelToUpdate);
-            
+
             if (result.IsError)
             {
                 return StatusCode(500, result.ErrorException.Message);
             }
-            
+
             return StatusCode(200);
         }
 
         // DELETE
         [HttpDelete("{channelId}")]
-        public ActionResult Delete(string channelId, [FromHeader]string userId)
+        public ActionResult Delete(string channelId, [FromHeader] string userId)
         {
-            // TODO
-            // Check auth - Unauthorised()
-            // Check permissions - Forbidden()
-            // Check channel exists - NotFound()
-            
             // Parse channel guid and user guid
             var channelGuid = ValidationUtil.ParseGuid(channelId);
             var userGuid = ValidationUtil.ParseGuid(userId);
@@ -128,14 +154,71 @@ namespace backend_api.Features.TaskForum.Controller
                 return BadRequest(Errors.InvalidGuid);
             }
 
+            // Check channel exists
+            var existError = CheckChannelExists(channelGuid);
+            if (existError != null)
+            {
+                return existError;
+            }
+
+            // Check user is an owner of the project so has permission to access
+            var task = _channelRepository.GetTask(channelGuid);
+            var error = CheckGroupMember(task.ReturnValue.Guid, userGuid);
+            if (error != null)
+            {
+                return error;
+            }
+
             var result = _channelRepository.Delete(channelGuid);
-            
+
             if (result.IsError)
             {
                 return StatusCode(500, result.ErrorException.Message);
             }
-            
+
             return StatusCode(200);
+        }
+
+        private ActionResult CheckGroupMember(Guid taskGuid, Guid userGuid)
+        {
+            var projectTaskRepository = new ProjectTaskRepository();
+            var project = projectTaskRepository.GetProject(taskGuid);
+            if (project.IsError)
+            {
+                return StatusCode(500, project.ErrorException.Message);
+            }
+
+            var projectRepository = new ProjectRepository();
+            var inGroup = projectRepository.IsGroupMember(userGuid, project.ReturnValue.Guid);
+            if (inGroup.IsError)
+            {
+                return StatusCode(500, inGroup.ErrorException.Message);
+            }
+
+            return !inGroup.ReturnValue ? StatusCode(403, Errors.NotGroupMember) : null;
+        }
+
+        private ActionResult CheckTaskExists(Guid taskGuid)
+        {
+            var projectTaskRepository = new ProjectTaskRepository();
+            var exists = projectTaskRepository.Exists(taskGuid);
+            if (exists.IsError)
+            {
+                return StatusCode(500, exists.ErrorException.Message);
+            }
+
+            return !exists.ReturnValue ? NotFound(Errors.TaskNotFound) : null;
+        }
+
+        private ActionResult CheckChannelExists(Guid channelGuid)
+        {
+            var exists = _channelRepository.Exists(channelGuid);
+            if (exists.IsError)
+            {
+                return StatusCode(500, exists.ErrorException.Message);
+            }
+
+            return !exists.ReturnValue ? StatusCode(404, Errors.ChannelNotFound) : null;
         }
     }
 }
